Clean project
- Router file change
- Fix top navigation menu condition
- Remove unused model files by now
- Start working on parameters component
