+++ frontend/src/components/color-classification/ColorGraph.vue
