+++ frontend/src/components/color-classification/ModelParameters.vue
@@ -1,0 +1,9 @@
+<script setup>
+</script>
+
+<template>
+    <section class="bg-black rounded-md p-4 w-full h-full z-10">
+        <p>Parameters</p>
+
+    </section>
+</template>
